+++ benchmarks/react/index.ts
@@ -1,3 +1,0 @@
-export * from './renderers/many-components';
-export * from './renderers/many-props';
-export * from './renderers/real-world';