+++ benchmarks/runner/lib/profile.js
@@ -1,0 +1,30 @@
+import fs from 'node:fs';
+import { Session } from 'node:inspector/promises';
+import { KitaJs } from './runners.js';
+import { generatePurchases } from './util.js';
+
+const purchases = generatePurchases(512_000);
+
+// Run to warm up the JIT
+for (let i = 0; i < 3; i++) {
+  KitaJs('Profile', purchases);
+}
+
+(async () => {
+  const session = new Session();
+
+  session.connect();
+
+  await session.post('Profiler.enable');
+  await session.post('Profiler.start');
+
+  // for (let i = 0; i < 10; i++) {
+  KitaJs('Profile', purchases);
+  // }
+
+  const { profile } = await session.post('Profiler.stop');
+
+  fs.writeFileSync('./profile.cpuprofile', JSON.stringify(profile));
+
+  // session.disconnect();
+})().catch(console.error);