+++ benchmarks/templates/index.tsx
@@ -1,0 +1,2 @@
+export * as Ghtml from './ghtml';
+export * as Normal from './normal';